+++ csrank/choicefunction/baseline.py
@@ -81,6 +81,3 @@ class AllPositive(ChoiceFunctions, Learner):
 
     def predict(self, X, **kwargs):
         return super().predict(X, **kwargs)
-
-    def set_tunable_parameters(self, **point):
-        pass