+++ csrank/discretechoice/baseline.py
@@ -33,6 +33,3 @@ class RandomBaselineDC(DiscreteObjectChooser, Learner):
 
     def predict(self, X, **kwargs):
         return super().predict(X, **kwargs)
-
-    def set_tunable_parameters(self, **point):
-        pass